requirements.txt
asgi_lifespan==2.1.0
cachetools==5.5.2
fastapi==0.115.12
httpx==0.28.1
passlib==1.7.4
pydantic==2.11.3
pydantic[email]
pydantic_settings==2.8.1
pytest==8.3.5
python_jose==3.3.0
SQLAlchemy==2.0.28
uvicorn==0.34.0
pymysql=1.1.1
bcrypt==4.1.2